+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - `axi_serializer`: serialize transactions with different IDs to the same ID.
 
 ### Changed
+- `axi_modify_address`: Simplify redundant `slv_resp_t` and `mst_resp_t` parameters to single
+  `resp_t` parameter.
 - `axi_modify_address_intf`:
   - Change name of slave port to `slv` and master port to `mst` and change name of associated
     parameters to align them with repository conventions.